requirements.txt
asgiref==3.8.1
babel==2.16.0
certifi==2024.8.30
charset-normalizer==3.4.0
dj-database-url==2.3.0
Django==4.2.16
django-countries==7.6.1
django-extra-views==0.14.0
django-haystack==3.3.0
django-oscar==3.2.4
django-phonenumber-field==6.4.0
django-tables2==2.3.4
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
factory-boy==3.2.1
Faker==30.8.0
idna==3.10
packaging==24.1
phonenumbers==8.13.48
pillow==11.0.0
psycopg2-binary==2.9.10
purl==1.6
pycountry==24.6.1
pysolr==3.10.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.7.0
